Replace FullDuplexStream+PairedStream with reuse of other streams
The PairedStream wasn't particularly optimized for netcoreapp2.1, but other streams we have implemented are. Rather than maintain so many independent streams, this removes the custom PairedStream implementation in favor of reusing our Pipe-as-Stream code, which already had most of the optimizations we need.

This introduces a requirement that FullDuplexStream paired streams must be flushed after writing before the reader can read it. This is a behavioral breaking change, but is consistent with other streams and actually makes this stream more useful in testing environments since it better emulates most other types of streams (e.g. if product code is failing to flush, that bug will now show up with the new FullDuplexStream behavior).