Avoid a subtly async Channel.dispose method
The public `Channel.dispose` method was typed as returning `void`, but the derived `ChannelClass.dispose` method would return a promise. As a result, public clients wouldn't know to `await` the result, and if `dispose` threw an exception, that would result in an unobserved promise rejection.

Since folks are not expected to await, rather than make the base method async, I made the derived method *not* async.