Fix stall in MultiplexingStream v3 protocol after receiving about 16 messages
The problem was we were creating *two* formatters on the stream. Each formatter piped the stream into its own decoder, but then only one loop was created to dequeue the formatter. As a result, each incoming frame went to *two* decoders, and only one was drained. The one that was never drained eventually fills up with messages to its high-water mark and then *pauses* the source stream, never to resume because no one drains the target stream.

The fix is to only create *one* formatter on the stream.

Fixes #578