Revert backward breaking change from v2.9
The new feature added in #536 that communicates channel failures to the remote party also caused channels to report failure if their owner `MultiplexingStream` was disposed of before the channel was. This broke several tests in the vs-servicehub repo and could theoretically break shipping code as well.
In this change I hide this particular behavioral change behind a setting that requires opt-in.